internal/cmd: Move code here from the main package
